requirements.txt
pandas==2.2.1
requests==2.32.3
requests-cache==1.2.0
SQLAlchemy==2.0.29
python-dotenv==1.0.1
psycopg2-binary==2.9.9
